% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getparams}
\alias{getparams}
\title{Retrieve List of API parameters}
\usage{
getparams(service = "")
}
\arguments{
\item{service}{character corresponding to API service (e.g. 'timeseries', 'network', etc)}
}
\value{
data.frame of paramters and descriptions
}
\description{
Retrieve List of API parameters
}
