% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mw}
\alias{mw}
\title{Retrieve Data from mesowest}
\usage{
mw(service, jsonsimplify = TRUE, ...)
}
\arguments{
\item{service}{character representing api service (e.g. 'metadata', 'timeseries')}

\item{jsonsimplify}{Should json results be compacted to a data.frame if possible (default = TRUE)}

\item{...}{parameters in the mesowest api corresponding to service (https://synopticlabs.org/api/explore/)}
}
\value{
List of data from JSON response
}
\description{
Retrieve Data from mesowest
}
\details{
If a request with nested lists (eg. climate variables per station) is made, the station metadata will be an awkward dataframe. It is suggested in this case to use jsonsimplify=FALSE.
}
\examples{
# Station metadata is a data.frame
mw( 'metadata', complete=1, stid=c('COOPTEEA3','mtmet'))
# Station metadata is a list
mw( 'metadata', complete=1,sensorvars=1, stid=c('COOPTEEA3','mtmet'))
}
