% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{mwtoken}
\alias{mwtoken}
\title{Get cached mesowest token}
\usage{
mwtoken()
}
\value{
cached token
}
\description{
Automatically searches the 'home' directory (~)  for a simple text file called ".mesowest" containing the token.
}
\seealso{
\code{\link{requestToken}}
}
