% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mwvariables}
\alias{mwvariables}
\title{Retrieve List of Station Variables}
\usage{
mwvariables()
}
\value{
nested list of potential station variables
}
\description{
Retrieve List of Station Variables
}
