% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{requestToken}
\alias{requestToken}
\title{Request an authentication token from mesowest}
\usage{
requestToken(apikey = "find")
}
\arguments{
\item{apikey}{string see details}
}
\description{
Request an authentication token from mesowest and cache locally
}
\details{
If apikey = 'find' (default), user prompted to interactively select the file containing the apikey. This file should be stored as a simple txt file with one line and no quotes. Otherwise, the function can take the apikey directly as a stirng.
}
\seealso{
\code{\link{mwtoken}}
}
